// src/requisitions/enums/requisition-type.ts
export enum RequisitionType {
    INTERNAL_TRANSFER = "INTERNAL_TRANSFER",
    RENT = "RENT",
    SALE = "SALE",
    CONSUMPTION = "CONSUMPTION",
    RETURN = "RETURN",
    ADJUSTMENT = "ADJUSTMENT",
    PURCHASE_RECEIPT = "PURCHASE_RECEIPT",
    //MAINTENACE = "MAINTENACE",
    //OUT_OF_SERVICE = "OUT_OF_SERVICE",
    TRANSFER = "TRANSFER",
}